Rework import page to use NavigationView
